pyproject.toml
[project]
name = "langgraph-multi-agent"
version = "0.1.0"
description = "Multi-agent workflow system using LangGraph, Ollama, and FastAPI"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=1.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "aiosqlite>=0.20.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg-pool>=3.3.0",
    "python-dotenv>=1.0.0",
    "langchain-tavily>=0.1.0",
    "langchain-community>=0.0.1",
    "chromadb>=0.5.0",
    "pypdf>=3.0.0",
    "langchain-chroma>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]
